Create a trait FixPerFileCollector and use it in SniffDecorator and FixerDecorator
