+++ phpinsights.php
@@ -3,6 +3,7 @@
 declare(strict_types=1);
 
 use NunoMaduro\PhpInsights\Domain\Insights\ForbiddenDefineGlobalConstants;
+use NunoMaduro\PhpInsights\Domain\Insights\ForbiddenTraits;
 use NunoMaduro\PhpInsights\Domain\Sniffs\ForbiddenSetterSniff;
 use PHP_CodeSniffer\Standards\Generic\Sniffs\CodeAnalysis\EmptyStatementSniff;
 use PHP_CodeSniffer\Standards\Generic\Sniffs\Files\LineLengthSniff;
@@ -95,6 +96,11 @@ return [
         EmptyStatementSniff::class => [
             'exclude' => [
                 'src/Domain/FileProcessors/FixerFileProcessor.php',
+            ],
+        ],
+        ForbiddenTraits::class => [
+            'exclude' => [
+                'src/Domain/Insights/FixPerFileCollector.php',
             ],
         ],
     ],
