+++ src/Domain/Insights/SniffDecorator.php
@@ -20,6 +20,8 @@ use PHP_CodeSniffer\Sniffs\Sniff;
  */
 final class SniffDecorator implements Sniff, Insight, HasDetails, Fixable
 {
+    use FixPerFileCollector;
+
     /**
      * @var \PHP_CodeSniffer\Sniffs\Sniff
      */
@@ -27,15 +29,6 @@ final class SniffDecorator implements Sniff, Insight, HasDetails, Fixable
 
     /** @var array<\NunoMaduro\PhpInsights\Domain\Details> */
     private $errors = [];
-
-    /**
-     * @var int
-     */
-    private $fixedCount = 0;
-    /**
-     * @var array<string, int>
-     */
-    private $fixPerFile = [];
 
     /**
      * @var array<string, \Symfony\Component\Finder\SplFileInfo>
@@ -143,46 +136,6 @@ final class SniffDecorator implements Sniff, Insight, HasDetails, Fixable
     public function addDetails(Details $details): void
     {
         $this->errors[] = $details;
-    }
-
-    public function incrementFix(): void
-    {
-        $this->fixedCount++;
-    }
-
-    public function addFileFixed(string $file): void
-    {
-        if (! \array_key_exists($file, $this->fixPerFile)) {
-            $this->fixPerFile[$file] = 0;
-        }
-
-        $this->fixPerFile[$file]++;
-        $this->incrementFix();
-    }
-
-    public function getTotalFix(): int
-    {
-        return $this->fixedCount;
-    }
-
-    /**
-     * @return array<Details>
-     */
-    public function getFixPerFile(): array
-    {
-        $details = [];
-        foreach ($this->fixPerFile as $file => $count) {
-            $message = 'issues fixed';
-            if ($count === 1) {
-                $message = 'issue fixed';
-            }
-
-            $details[] = (new Details())
-                ->setMessage(sprintf('%s %s', $count, $message))
-                ->setFile($file);
-        }
-
-        return $details;
     }
 
     private function skipFilesFromIgnoreFiles(InsightFile $file): bool
